Added wrapper script for git for sudo
This will be found first when executing sudo git and checks whether
it is running with super user privileges and if so refuses to do
that and instead just prints a friendly message to please run git
without sudo.

Reason: Apparently users tend to do a "sudo git pull" when updating
OctoPrint, effectively screwing up their permissions in the process
and then running into problems later on. This should hopefully act
as an educational measure to stop this from happening.